day 2
Make some change interpolation didn't work for polynomials of degree n for n+1 datapoints because of slope function which is too high. Tmrw i will think about some change like polynomials of degee n+1 for n+1 datapoints which are infinite so i have to think of a new "equation" for datapoints to make the slope of the function to not be that steep. Or if i don't find a good one maybe going to do some lecture like introduction on statistical learningL